Refactor: Home screen and service hooks
