+++ src/navigators/StackNavigator.tsx
@@ -4,7 +4,7 @@ import {
   createNativeStackNavigator,
 } from '@react-navigation/native-stack';
 
-import Home from '../screens/Home';
+import Home from '../screens/home/Home';
 import MovieDetails from '../screens/MovieDetails';
 import Search from '../screens/Search';
 import {useThemeColors} from '../theme/theme';
